Use HTTPS URL for the dmz submodule
The git URL requires GitHub SSH keys be set up on the each developer's
machine, and also on the CI machines. As an alternative, we fork the
upstream repository and use an HTTPS URL for the submodule -- public
HTTPS URLs should work even with no credentials have been set up on
the machine.